minor refactoring and tidying up
