+++ src/main/java/uk/gov/ons/ctp/response/sample/config/ThreadPoolConfig.java
@@ -4,6 +4,7 @@ import org.springframework.aop.interceptor.AsyncUncaughtExceptionHandler;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.scheduling.annotation.AsyncConfigurer;
 import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
+import uk.gov.ons.ctp.response.sample.exception.AsyncSampleUncaughtExceptionHandler;
 
 import java.util.concurrent.Executor;
